split to two separate validations
