+++ services/ui-src/src/measures/2024/shared/globalValidations/index.ts
@@ -7,7 +7,7 @@ export * from "shared/globalValidations/validateAtLeastOneDefinitionOfPopulation
 export * from "shared/globalValidations/validateHybridMeasurePopulation";
 export * from "./validateAtLeastOneDataSourceType";
 export * from "shared/globalValidations/validateAtLeastOneDeliverySystem";
-export * from "./validateAtLeastOneDeviationFieldFilled";
+export * from "shared/globalValidations/validateDeviationTextFieldFilled";
 export * from "shared/globalValidations/validateAtLeastOneRateComplete";
 export * from "shared/globalValidations/validateBothDatesInRange";
 export * from "shared/globalValidations/validateDateRangeRadioButtonCompletion";
